proof-params: move check to build time
This way there isn't a delay while we wait for parameters to download
before trying to send a transaction
